separate StyledIcon into its own file to stay DRY
